view statistic

+++ src/main/java/org/frog/controller/mentor/DeleteScheduleController.java
@@ -5,8 +5,10 @@ import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.frog.DAO.ScheduleDAO;
+import org.frog.DAO.SlotDao;
 import org.frog.controller.auth.AuthenticationServlet;
 import org.frog.model.Account;
+import org.frog.model.Slot;
 import org.frog.utility.DateTimeHelper;
 
 import java.io.IOException;
@@ -25,7 +27,15 @@ public class DeleteScheduleController extends AuthenticationServlet {
             if(slotID!=null){
                 String [] infoSlotID = slotID.split("_");
                 ScheduleDAO scheduleDAO = new ScheduleDAO();
-                scheduleDAO.deleteDayFreeByMentor(account.getId(), DateTimeHelper.convertStringToDateByDay(infoSlotID[0]), Integer.parseInt(infoSlotID[1]));
+                SlotDao slDAO = new SlotDao();
+                Slot id = slDAO.getTimeSlot(Integer.parseInt(infoSlotID[1]));
+                if(DateTimeHelper.compareDayIDtoNow(infoSlotID[0], id.getStart_at(), id.getEnd_at())){
+                    scheduleDAO.deleteDayFreeByMentor(account.getId(), DateTimeHelper.convertStringToDateByDay(infoSlotID[0]), Integer.parseInt(infoSlotID[1]));
+
+                }
+                else{
+                    req.getSession().setAttribute("DeleteSlotError", "Can not delete slot in passed");
+                }
                 resp.sendRedirect("/Frog/mentor/schedule?today="+day);
             }
         }catch(Exception e){
